Move get path iter to pathfinding
