+++ indra/explanation/model_checker/__init__.py
@@ -1,4 +1,5 @@
-from .model_checker import ModelChecker, PathResult, PathMetric, get_path_iter
+from .model_checker import ModelChecker, PathResult, PathMetric
+from ..pathfinding import get_path_iter
 from .pysb import PysbModelChecker
 from .signed_graph import SignedGraphModelChecker
 from .unsigned_graph import UnsignedGraphModelChecker
